requirements.txt
sqlalchemy==2.0.25
psycopg[binary]>=3.1.0
pandas==2.1.4
streamlit==1.29.0
python-dotenv==1.0.0
